add interface for NodeJS libuv event loop to poll/select generic socket file descriptors
NodeJS provides only limited support to poll/select file descriptors. For
example, it is not possible to poll file descriptors of UDP sockets. This
commit adds an interface to the internal libuv event loop of NodeJS which
enables adding a generic file descriptor to the event loop. Please note
that some plattforms still limit the kind of file descriptors that can be
used with libuv, e.g. only socket file descriptors can be used under
Windows.